requirements.txt
cerebras-appliance==2.9.0
cerebras-sdk==2.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
dgl==2.1.0
dill==0.4.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
grpcio==1.51.1
grpcio-tools==1.51.1
h5py==3.13.0
hdf5plugin==5.1.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
packaging==26.0
pandas==2.2.2
pillow==12.2.0
protobuf==4.21.12
psutil==7.1.3
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.33.1
scipy==1.17.1
six==1.17.0
sympy==1.14.0
tblib==3.0.0
torch==2.11.0+cpu
torchdata==0.11.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
zstandard==0.25.0
